requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
scikit-learn
pandas
numpy
shap
joblib
psycopg2-binary
python-dotenv